+++ src/api/environments.ts
@@ -14,7 +14,7 @@ import {
 import { BASE_PATH } from '../openapi/base';
 
 import { BasePermitApi, IPagination, PermitApiError } from './base'; // eslint-disable-line @typescript-eslint/no-unused-vars
-import { ApiContext, ApiKeyLevel, PermitContextError } from './context'; // eslint-disable-line @typescript-eslint/no-unused-vars
+import { ApiContext, ApiContextLevel, ApiKeyLevel, PermitContextError } from './context'; // eslint-disable-line @typescript-eslint/no-unused-vars
 
 export {
   APIKeyRead,
@@ -197,7 +197,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
    */
   public async list(params: IListEnvironments): Promise<EnvironmentRead[]> {
     const { projectKey, page = 1, perPage = 100 } = params;
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.listEnvironments({
@@ -221,7 +222,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
    * @throws {@link PermitContextError} If the configured {@link ApiContext} does not match the required endpoint context.
    */
   public async get(projectKey: string, environmentKey: string): Promise<EnvironmentRead> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.getEnvironment({
@@ -272,7 +274,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
    * @throws {@link PermitContextError} If the configured {@link ApiContext} does not match the required endpoint context.
    */
   public async getStats(projectKey: string, environmentKey: string): Promise<EnvironmentStats> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.statsEnvironments({
@@ -296,7 +299,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
    * @throws {@link PermitContextError} If the configured {@link ApiContext} does not match the required endpoint context.
    */
   public async getApiKey(projectKey: string, environmentKey: string): Promise<APIKeyRead> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.apiKeys.getEnvironmentApiKey({
@@ -322,7 +326,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
     projectKey: string,
     environmentData: EnvironmentCreate,
   ): Promise<EnvironmentRead> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.createEnvironment({
@@ -350,7 +355,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
     environmentKey: string,
     environmentData: EnvironmentUpdate,
   ): Promise<EnvironmentRead> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.updateEnvironment({
@@ -384,7 +390,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
     environmentKey: string,
     copyParams: EnvironmentCopy,
   ): Promise<EnvironmentRead> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       return (
         await this.environments.copyEnvironment({
@@ -408,7 +415,8 @@ export class EnvironmentsApi extends BasePermitApi implements IEnvironmentsApi {
    * @throws {@link PermitContextError} If the configured {@link ApiContext} does not match the required endpoint context.
    */
   public async delete(projectKey: string, environmentKey: string): Promise<void> {
-    await this.ensureContext(ApiKeyLevel.PROJECT_LEVEL_API_KEY);
+    await this.ensureAccessLevel(ApiKeyLevel.ENVIRONMENT_LEVEL_API_KEY);
+    await this.ensureContext(ApiContextLevel.ORGANIZATION);
     try {
       await this.environments.deleteEnvironment({
         projId: projectKey,
